+++ homeassistant/components/homematicip_cloud/const.py
@@ -1,0 +1,11 @@
+"""Constants for the HomematicIP Cloud component."""
+import logging
+
+_LOGGER = logging.getLogger('homeassistant.components.homematicip_cloud')
+
+DOMAIN = 'homematicip_cloud'
+
+CONF_NAME = 'name'
+CONF_ACCESSPOINT = 'accesspoint'
+CONF_AUTHTOKEN = 'authtoken'
+CONF_PIN = 'pin'
